moved related components to DependencyManagementPane
