--- src/components/DependencyInstalling/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default } from './DependencyInstalling';
