+++ src/components/DependencyManagementPane/AddDependencySearchBox.js
