+++ src/components/DependencyManagementPane/AddDependencySearchProvider.js
