+++ src/components/DependencyManagementPane/AddDependencySearchResult.js
