+++ src/components/DependencyManagementPane/AlgoliaLogo.js
