+++ src/components/DependencyManagementPane/DependencyDetails.js
