+++ src/components/DependencyManagementPane/DependencyInstalling.js
